Added safety prompts when turning edit lock off for single or multiple cells. Edit widget now saves data when the enter key is pressed.
git-svn-id: http://subversion.wr.usgs.gov/repos/prog/isis3/trunk@2944 41f8697f-d340-4b68-9986-7bafba869bb8

+++ isis/src/qisis/apps/cneteditor/AbstractPointItem.cpp
@@ -312,6 +312,12 @@ namespace Isis
   {
     if (point)
     {
+      if (point->IsEditLocked()) {
+        iString msg = "Point [" + getData() + "] is edit locked and cannot be "
+            "deleted";
+        throw iException::Message(iException::User, msg, _FILEINFO_);
+      }
+
       ControlPoint * tempPoint = point;
       point = NULL;
       tempPoint->Parent()->DeletePoint(tempPoint);
